release: v0.5.1 :tada:
